+++ CHANGES.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## V0.7.7
+
+Added a new action actions:
+* `vsphere.host_get` - retrieves the Summary information for an ESX host.
+* `vsphere.host_network_hints_get` - retrieves the Network Hints for an ESX host.
+* `vsphere.vm_guest_info_get` - retrieves the Guest information for a VM.
+* `vsphere.vm_runtime_info_get` - retrieves the Runtime information for a VM.
+
+Contributed by Nick Maludy (Encore Technologies).
+
 ## V0.7.6
 
 Minor linting fixes